% Generated by roxygen2 (4.0.0): do not edit by hand
\name{arm}
\alias{arm}
\title{Arm a value with a function which fires when evaluated.}
\usage{
arm(expr, fun)
}
\arguments{
\item{expr}{an expression}

\item{fun}{a function of one argument.}
}
\value{
an expression which when evaluated returns the result of expr,
after evaluating fun(expr).
}
\description{
Arm a value with a function which fires when evaluated.
}
\details{
Thanks very much to Stefan Milton Bache for providing the code to replace the
now defunct \code{pipe_with} function which uses \code{arm}, which is necessary for \code{\%W>\%}
to function. Retrieved from this gist: \url{https://gist.github.com/smbache/86fe703fa46e39df33ea}
}

