% Generated by roxygen2 (4.0.0): do not edit by hand
\name{\%W>\%}
\alias{\%W>\%}
\title{Pipe/warp an object downwards into a function call/expression/dungeon level.}
\usage{
lhs \%W>\% rhs
}
\arguments{
\item{lhs}{That which is to be warped}

\item{rhs}{the function/dungeon level to be warped into}
}
\description{
Also known as the "warp" operator, \code{\%W>\%} does exactly the same thing as \code{\link{\%>\%}}:
it pipes the left-hand side into an expression on the right-hand side. But with
a bonus: when you use \code{\%W>\%} you hear the sound that --- if you are anything
like me --- goes through your head every time you use a pipe, only for real!
}
\examples{
seq(1,10,0.1) \%W>\% sin \%T>\% print \%T>\% plot(type = 'l') \%T>\% (l(x ~ Sys.sleep(1))) \%W>\%
sin \%T>\% print \%T>\% plot(type = 'l') \%T>\% (l(x ~ Sys.sleep(1))) \%W>\%
sin \%T>\% print \%T>\% plot(type = 'l')

\dontrun{

library(dplyr)
library(Lahman)
Batting \%W>\%
group_by(playerID) \%T>\% print \%T>\% (l(x ~ Sys.sleep(2))) \%W>\%
summarise(total = sum(G)) \%T>\% print \%T>\% (l(x ~ Sys.sleep(2))) \%W>\%
arrange(desc(total))

}
}

